backend/requirements.txt
fastapi==0.95.0
uvicorn==0.21.1
pydantic==1.10.7
python-multipart==0.0.6
python-jose==3.3.0
passlib==1.7.4
psycopg2-binary==2.9.6
geopandas==0.12.2
rasterio==1.3.6
shapely==2.0.1
fiona==1.9.3
earthengine-api==0.1.341
sentinel-hub==3.8.1
gdal==3.6.4
numpy==1.24.3
pandas==2.0.1
matplotlib==3.7.1
PyPDF2==3.0.1
reportlab==3.6.13